Split galloc crate and redo all examples
