+++ examples/README.md
@@ -49,7 +49,7 @@ use gstd::{ext, msg};
 
 #[no_mangle]
 pub unsafe extern "C" fn handle() {
-    let new_msg = String::from_utf8(msg::load()).expect("Invalid message: should be utf-8");
+    let new_msg = String::from_utf8(msg::load_bytes()).expect("Invalid message: should be utf-8");
 
     if &new_msg == "PING" {
         msg::send(msg::source(), b"PONG", 10_000_000);
