+++ examples/async/Cargo.toml
@@ -9,4 +9,5 @@ crate-type = ["cdylib"]
 
 [dependencies]
 gcore = { path = "../../gcore", features = ["debug"] }
+gstd = { path = "../../gstd", features = ["debug"] }
 gstd-async = { path = "../../gstd-async" }
