+++ examples/collector/src/lib.rs
@@ -4,7 +4,8 @@
 extern crate alloc;
 
 use alloc::collections::BTreeMap;
-use gcore::{msg, prelude::*};
+use gstd::prelude::*;
+use gcore::msg;
 
 static mut MY_COLLECTION: BTreeMap<usize, String> = BTreeMap::new();
 
@@ -12,7 +13,7 @@ static mut COUNTER: usize = 0;
 
 #[no_mangle]
 pub unsafe extern "C" fn handle() {
-    let new_msg = String::from_utf8(msg::load()).expect("Invalid message: should be utf-8");
+    let new_msg = String::from_utf8(gstd::msg::load_bytes()).expect("Invalid message: should be utf-8");
     if new_msg == "log" {
         let collapsed = mem::replace(&mut MY_COLLECTION, BTreeMap::new())
             .into_iter()
