+++ examples/mem/Cargo.toml
@@ -10,3 +10,4 @@ crate-type = ["cdylib"]
 
 [dependencies]
 gcore = { path = "../../gcore", features = ["debug"] }
+gstd = { path = "../../gstd", features = ["debug"] }
