+++ examples/panicker/src/lib.rs
@@ -1,8 +1,6 @@
 #![no_std]
 #![feature(default_alloc_error_handler)]
 
-use gcore::prelude::*;
-
 #[no_mangle]
 pub unsafe extern "C" fn handle() {
     panic!("I just panic every time")
