+++ examples/ping/src/lib.rs
@@ -1,13 +1,14 @@
 #![no_std]
 #![feature(default_alloc_error_handler)]
 
-use gcore::{ext, msg, prelude::*};
+use gcore::{ext, msg};
+use gstd::prelude::*;
 
 static mut MESSAGE_LOG: Vec<String> = vec![];
 
 #[no_mangle]
 pub unsafe extern "C" fn handle() {
-    let new_msg = String::from_utf8(msg::load()).expect("Invalid message: should be utf-8");
+    let new_msg = String::from_utf8(gstd::msg::load_bytes()).expect("Invalid message: should be utf-8");
 
     if new_msg == "PING" {
         msg::reply(b"PONG", 10_000_000, 0);
