+++ galloc/README.md
@@ -1,0 +1,3 @@
+# gcore
+
+Core library for Gear smart contracts.
