+++ gcore/src/lib.rs
@@ -19,9 +19,7 @@
 #![no_std]
 #![cfg_attr(feature = "strict", deny(warnings))]
 
-mod bail;
 pub mod msg;
-pub mod prelude;
 
 mod general;
 pub use general::*;
@@ -29,3 +27,7 @@ pub use general::*;
 mod utils;
 #[cfg(feature = "debug")]
 pub use utils::ext;
+
+pub mod prelude {
+    pub use core::prelude::*;
+}
