+++ gstd/Cargo.toml
@@ -7,6 +7,7 @@ license = "GPL-3.0"
 
 [dependencies]
 gcore = { path = "../gcore" }
+galloc = { path = "../galloc" }
 codec = { package = "parity-scale-codec", version = "2.0.0", default-features = false }
 
 [features]
